Update strukture dashboard_pelanggan
