+++ config/database.php
@@ -1,0 +1,18 @@
+<?php
+$host = 'localhost';
+$dbname = 'Pare_sports';  // Diubah ke nama database baru
+$username = 'root';
+$password = '';
+
+try {
+    $conn = new mysqli($host, $username, $password, $dbname);
+    
+    if ($conn->connect_error) {
+        die("Connection failed: " . $conn->connect_error);
+    }
+    
+    $conn->set_charset("utf8mb4");
+} catch (Exception $e) {
+    die("Database connection failed: " . $e->getMessage());
+}
+?>
