+++ includes/footer.php
@@ -1,0 +1,22 @@
+<footer class="footer py-4 bg-dark text-white">
+  <div class="container">
+    <div class="row">
+      <div class="col-md-4">
+        <h6>PARESPORTS</h6>
+        <p>Jl. Pemuda, Parepare, Sulawesi Selatan</p>
+      </div>
+      <div class="col-md-4">
+        <h6>Perusahaan</h6>
+        <p><a href="#" class="text-white-50">Tentang Kami</a></p>
+        <p><a href="#" class="text-white-50">Ekosistem</a></p>
+      </div>
+      <div class="col-md-4">
+        <h6>Hubungi Kami</h6>
+        <p><a href="#" class="text-white-50">Kontak</a></p>
+      </div>
+    </div>
+    <div class="text-center mt-3">
+      <small class="text-white-50">© <?= date('Y') ?> PARESPORTS. All Rights Reserved.</small>
+    </div>
+  </div>
+</footer>
