+++ proses/proses_register.php
@@ -1,0 +1,50 @@
+<?php
+session_start();
+require_once '../config/database.php';
+
+if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+    $nama = $_POST['nama'];
+    $email = $_POST['email'];
+    $password = $_POST['password'];
+    $confirm_password = $_POST['confirm_password'];
+    
+    if ($password !== $confirm_password) {
+        $_SESSION['register_error'] = "Password dan konfirmasi password tidak cocok!";
+        header('Location: ../index.php');
+        exit();
+    }
+    
+    $hashed_password = password_hash($password, PASSWORD_DEFAULT);
+    
+    // Cek apakah email sudah terdaftar
+    $query = "SELECT * FROM users WHERE email = ?";
+    $stmt = $conn->prepare($query);
+    $stmt->bind_param('s', $email);
+    $stmt->execute();
+    $result = $stmt->get_result();
+    
+    if ($result->num_rows > 0) {
+        $_SESSION['register_error'] = "Email sudah terdaftar!";
+        header('Location: ../index.php');
+        exit();
+    }
+    
+    // Insert user baru
+    $query = "INSERT INTO users (nama, email, password, role) VALUES (?, ?, ?, 'user')";
+    $stmt = $conn->prepare($query);
+    $stmt->bind_param('sss', $nama, $email, $hashed_password);
+    
+    if ($stmt->execute()) {
+        $_SESSION['register_success'] = "Pendaftaran berhasil! Silakan login.";
+        header('Location: ../index.php');
+        exit();
+    } else {
+        $_SESSION['register_error'] = "Pendaftaran gagal. Silakan coba lagi!";
+        header('Location: ../index.php');
+        exit();
+    }
+} else {
+    header('Location: ../index.php');
+    exit();
+}
+?>
